+++ fiftyone/labs/schema.py
@@ -8,7 +8,7 @@ FiftyOne Teams schema.
 import strawberry as gql
 
 from .extensions import EndSession
-from .mutations import Mutation
-from .queries import Query
+from .mutation import Mutation
+from .query import Query
 
 schema = gql.Schema(query=Query, mutation=Mutation, extensions=[EndSession])