+++ fiftyone/utils/data/importers.py
@@ -676,7 +676,11 @@ class ImportPathsMixin(object):
                 )
 
             data_map = fos.read_json(data_path)
-            return {to_uuid(k): v for k, v in data_map.items()}
+            data_root = os.path.dirname(data_path)
+            return {
+                to_uuid(k): os.path.join(data_root, v)
+                for k, v in data_map.items()
+            }
 
         if not fos.isdir(data_path):
             raise ValueError("Data directory '%s' does not exist" % data_path)
@@ -2069,6 +2073,8 @@ class ImageClassificationDirectoryTreeImporter(LabeledImageDatasetImporter):
         compute_metadata (False): whether to produce
             :class:`fiftyone.core.metadata.ImageMetadata` instances for each
             image when importing
+        classes (None): an optional string or list of strings specifying a
+            subset of classes to load
         unlabeled ("_unlabeled"): the name of the subdirectory containing
             unlabeled images
         shuffle (False): whether to randomly shuffle the order in which the
@@ -2082,11 +2088,14 @@ class ImageClassificationDirectoryTreeImporter(LabeledImageDatasetImporter):
         self,
         dataset_dir,
         compute_metadata=False,
+        classes=None,
         unlabeled="_unlabeled",
         shuffle=False,
         seed=None,
         max_samples=None,
     ):
+        classes = _to_list(classes)
+
         super().__init__(
             dataset_dir=dataset_dir,
             shuffle=shuffle,
@@ -2095,6 +2104,7 @@ class ImageClassificationDirectoryTreeImporter(LabeledImageDatasetImporter):
         )
 
         self.compute_metadata = compute_metadata
+        self.classes = classes
         self.unlabeled = unlabeled
 
         self._classes = None
@@ -2139,6 +2149,7 @@ class ImageClassificationDirectoryTreeImporter(LabeledImageDatasetImporter):
         samples = []
         classes = set()
         sep = fos.sep(self.dataset_dir)
+        whitelist = set(self.classes) if self.classes is not None else None
 
         for relpath in fos.list_files(self.dataset_dir, recursive=True):
             chunks = relpath.split(sep, 1)
@@ -2149,6 +2160,9 @@ class ImageClassificationDirectoryTreeImporter(LabeledImageDatasetImporter):
             if label.startswith("."):
                 continue
 
+            if whitelist is not None and label not in whitelist:
+                continue
+
             if label == self.unlabeled:
                 label = None
             else:
@@ -2157,13 +2171,17 @@ class ImageClassificationDirectoryTreeImporter(LabeledImageDatasetImporter):
             path = fos.join(self.dataset_dir, relpath)
             samples.append((path, label))
 
-        classes = sorted(classes)
         samples = self._preprocess_list(samples)
 
         if self.compute_metadata:
             metadata_map = self._get_remote_metadata([s[0] for s in samples])
         else:
             metadata_map = {}
+
+        if whitelist is not None:
+            classes = self.classes
+        else:
+            classes = sorted(classes)
 
         self._samples = samples
         self._metadata_map = metadata_map
@@ -2184,6 +2202,16 @@ class ImageClassificationDirectoryTreeImporter(LabeledImageDatasetImporter):
         return len(fos.list_files(dataset_dir, recursive=True))
 
 
+def _to_list(arg):
+    if arg is None:
+        return None
+
+    if etau.is_container(arg):
+        return list(arg)
+
+    return [arg]
+
+
 class VideoClassificationDirectoryTreeImporter(LabeledVideoDatasetImporter):
     """Importer for a viideo classification directory tree stored on disk.
 
@@ -2195,6 +2223,8 @@ class VideoClassificationDirectoryTreeImporter(LabeledVideoDatasetImporter):
         compute_metadata (False): whether to produce
             :class:`fiftyone.core.metadata.VideoMetadata` instances for each
             video when importing
+        classes (None): an optional string or list of strings specifying a
+            subset of classes to load
         unlabeled ("_unlabeled"): the name of the subdirectory containing
             unlabeled images
         shuffle (False): whether to randomly shuffle the order in which the
@@ -2208,11 +2238,14 @@ class VideoClassificationDirectoryTreeImporter(LabeledVideoDatasetImporter):
         self,
         dataset_dir,
         compute_metadata=False,
+        classes=None,
         unlabeled="_unlabeled",
         shuffle=False,
         seed=None,
         max_samples=None,
     ):
+        classes = _to_list(classes)
+
         super().__init__(
             dataset_dir=dataset_dir,
             shuffle=shuffle,
@@ -2221,6 +2254,7 @@ class VideoClassificationDirectoryTreeImporter(LabeledVideoDatasetImporter):
         )
 
         self.compute_metadata = compute_metadata
+        self.classes = classes
         self.unlabeled = unlabeled
 
         self._classes = None
@@ -2269,6 +2303,7 @@ class VideoClassificationDirectoryTreeImporter(LabeledVideoDatasetImporter):
         samples = []
         classes = set()
         sep = fos.sep(self.dataset_dir)
+        whitelist = set(self.classes) if self.classes is not None else None
 
         for relpath in fos.list_files(self.dataset_dir, recursive=True):
             chunks = relpath.split(sep, 1)
@@ -2277,6 +2312,9 @@ class VideoClassificationDirectoryTreeImporter(LabeledVideoDatasetImporter):
 
             label = chunks[0]
             if label.startswith("."):
+                continue
+
+            if whitelist is not None and label not in whitelist:
                 continue
 
             if label == self.unlabeled:
@@ -2288,7 +2326,11 @@ class VideoClassificationDirectoryTreeImporter(LabeledVideoDatasetImporter):
             samples.append((path, label))
 
         samples = self._preprocess_list(samples)
-        classes = sorted(classes)
+
+        if whitelist is not None:
+            classes = self.classes
+        else:
+            classes = sorted(classes)
 
         if self.compute_metadata:
             metadata_map = self._get_remote_metadata([s[0] for s in samples])